refactor(api): migrate ValueType parameters to object pattern
Complete the NumPy-aligned API migration by replacing all `ValueType`
parameters with `object` in the public API surface. This matches NumPy's
`PyArray_FromAny` behavior where any input is accepted and converted
via `np.asanyarray()`.

Public API Changes:
- np.power(arr, object) - scalar/array-like exponent
- np.floor_divide(arr, object) - scalar/array-like divisor
- np.left_shift(arr, object) - scalar/array-like shift
- np.right_shift(arr, object) - scalar/array-like shift
- np.full(object, ...) - scalar fill value (10 overloads)
- np.equal/not_equal/less/greater/etc. - object pattern (12 overloads)
- NDArray.itemset(object) - scalar value to set
- NDArray.SetValue(object) - removed redundant ValueType overload
- UnmanagedStorage.SetAtIndexUnsafe(object) - scalar value

NumPy 2.x Alignment:
- Added NDArray.item() method - the NumPy 2.x replacement for np.asscalar
  - item() - extract scalar from size-1 arrays
  - item(index) - flat indexing with negative index support
  - item(i, j) / item(i, j, k) - multi-dimensional indexing
  - item<T>() - type-converting variant
- Marked np.asscalar as [Obsolete] (removed in NumPy 2.0)

Internal Changes:
- TensorEngine: removed ValueType overloads for Power, FloorDivide,
  Clip, LeftShift, RightShift (6 methods)
- DefaultEngine: removed corresponding override implementations

Fixes:
- NumSharp.Bitmap: == null -> is null (2 occurrences)
- examples/NeuralNetwork: == null -> is null
- Tests: explicit casts for NDArray->scalar (3 files)
- np.random.gamma: removed explicit ValueType cast

+++ test/NumSharp.UnitTest/Math/np.math.Test.cs
@@ -43,7 +43,7 @@ namespace NumSharp.UnitTest.Maths
         {
             var data = new int[,] {{0, 1}, {0, 5}};
 
-            int s = np.sum(data);
+            int s = (int)np.sum(data);
             Assert.AreEqual(s, 6);
 
             var s0 = np.sum(data, axis: 0);
